[feat] add profile-stats
