+++ db/initCategories.js
